Algorithms

Available Algorithms

This section of the manual describes the algorithms that are available in this package.

User interface

Posing and solving a reachability problem

Continuous-time reachability

Decomposition-based approach

Reachability.ReachSets.BFFPSV18 — Type.

BFFPSV18 <: ContinuousPost

Implementation of the reachability algorithm for purely continuous linear time-invariant systems using block decompositons by S. Bogomolov, M. Forets, G. Frehse, A. Podelski, C. Schilling and F. Viry [1].

Fields

options – an Options structure that holds the algorithm-specific options

Notes

The following options are available:

option :approx_model of type String has default value 'forward'; model for bloating/continuous time analysis
option :algorithm of type String has default value 'explicit'; algorithm backend
option :δ of type Float64 with alias :sampling_time has default value '0.01'; time step
option :vars of type AbstractArray{Int64,1} has default value 'Int64[]'; variables of interest; default: all variables
option :partition of type AbstractArray{#s143,1} where #s143<:AbstractArray{Int64,1} has default value 'Array{Int64,1}[[]]'; block partition; a block is represented by a vector containing its indices
option :lazy_sih of type Bool has default value 'false'; use a lazy symmetric interval hull in discretization?
option :lazy_expm of type Bool has default value 'false'; use a lazy matrix exponential all the time?
option :lazy_expm_discretize of type Bool has default value 'false'; use a lazy matrix exponential in discretization?
option :pade_expm of type Bool has default value 'false'; use the Padé approximant method (instead of Julia's  built-in 'exp') to compute the lazy matrix exponential in discretization?
option :assume_sparse of type Bool has default value 'false'; use an analysis for sparse discretized matrices?
option :lazy_X0 of type Bool has default value 'false'; keep the discretized and decomposed initial states a lazy set?
option :lazy_inputs_interval of type Union{Int64, Function} has default value 'getfield(Reachability.ReachSets, Symbol("##23#24"))()'; length of interval in which the inputs are handled as a lazy set (``-1`` for 'never'); may generally also be a predicate over indices; the default corresponds to ``-1``
option :block_types of type Union{Nothing, Dict{Type{#s45} where #s45<:LazySet,AbstractArray{#s44,1} where #s44<:AbstractArray{Int64,1}}} has default value 'nothing'; short hand to set ':block_types_init' and ':block_types_iter'
option :block_types_init of type Union{Nothing, Dict{Type{#s45} where #s45<:LazySet,AbstractArray{#s44,1} where #s44<:AbstractArray{Int64,1}}} has default value 'nothing'; set type for the approximation of the initial states for each block
option :block_types_iter of type Union{Nothing, Dict{Type{#s45} where #s45<:LazySet,AbstractArray{#s44,1} where #s44<:AbstractArray{Int64,1}}} has default value 'nothing'; set type for the approximation of the states ``X_k``, ``k>0``, for each block
option :ε of type Float64 has default value 'Inf'; short hand to set `:ε_init` and `:ε_iter`
option :ε_init of type Float64 has default value 'Inf'; error bound for the approximation of the initial states(during decomposition)
option :ε_iter of type Float64 has default value 'Inf'; error bound for the approximation of the states ``X_k``, ``k>0``
option :set_type of type Union{Type{HPolygon}, Type{Hyperrectangle}, Type{Interval}} has default value 'LazySets.Hyperrectangle'; short hand to set `:set_type_init` and `:set_type_iter`
option :set_type_init of type Union{Type{HPolygon}, Type{Hyperrectangle}, Type{Interval}} has default value 'LazySets.Hyperrectangle'; set type for the approximation of the initial states(during decomposition)
option :set_type_iter of type Union{Type{HPolygon}, Type{Hyperrectangle}, Type{Interval}} has default value 'LazySets.Hyperrectangle'; set type for the approximation of the states ``X_k``, ``k>0``
option :template_directions of type Symbol has default value 'nothing'; short hand to set `template_directions_init` and `template_directions_iter`
option :template_directions_init of type Symbol has default value 'nothing'; directions to use for the approximation of the initial states (during decomposition)
option :template_directions_iter of type Symbol has default value 'nothing'; directions to use for the approximation of the states ``X_k``, ``k>0``, for each block
option :assume_homogeneous of type Bool has default value 'false'; ignore dynamic inputs during the analysis?
option :eager_checking of type Bool has default value 'true'; terminate as soon as property violation was detected?

Algorithm

We refer to [1] for technical details.

[1] Reach Set Approximation through Decomposition with Low-dimensional Sets and High-dimensional Matrices. S. Bogomolov, M. Forets, G. Frehse, A. Podelski, C. Schilling, F. Viry. HSCC '18 Proceedings of the 21st International Conference on Hybrid Systems: Computation and Control (part of CPS Week).

source