Discretization

Discretize

Discretize

Dense-time reachability

Discrete-time reachability

Reachability.ReachSets.discretize — Function.

discretize(cont_sys, δ; [approx_model], [pade_expm], [lazy_expm], [lazy_sih])

Discretize a continuous system of ODEs with nondeterministic inputs.

Input

cont_sys – continuous system

δ – step size

approx_model – the method to compute the approximation model for the discretization, among:

forward – use forward-time interpolation

backward – use backward-time interpolation

firstorder – use first order approximation of the ODE

nobloating – do not bloat the initial states (use for discrete-time reachability)

pade_expm – (optional, default = false) if true, use Pade approximant method to compute matrix exponentials of sparse matrices; otherwise use Julia's buil-in expm

lazy_expm – (optional, default = false) if true, compute the matrix exponential in a lazy way (suitable for very large systems)

lazy_sih – (optional, default = true) if true, compute the symmetric interval hull in a lazy way (suitable if only a few dimensions are of interest)

Output

A discrete system.

Notes

This function applies an approximation model to transform a continuous affine system into a discrete affine system. This transformation allows to do dense time reachability, i.e. such that the trajectories of the given continuous system are included in the computed flowpipe of the discretized system. For discrete-time reachability, use approx_model="nobloating".

source

Dense-time reachability

Reachability.ReachSets.discr_bloat_firstorder — Function.

bloat_firstorder(cont_sys, δ)

Compute bloating factors using first order approximation.

Input

cont_sys – a continuous affine system

δ – step size

Notes

In this algorithm, the infinity norm is used. See also: discr_bloat_interpolation for more accurate (less conservative) bounds.

Algorithm

This uses a first order approximation of the ODE, and matrix norm upper bounds, see Le Guernic, C., & Girard, A., 2010, Reachability analysis of linear systems using support functions. Nonlinear Analysis: Hybrid Systems, 4(2), 250-262.

source

Reachability.ReachSets.discr_bloat_interpolation — Function.

discr_bloat_interpolation(cont_sys, δ, approx_model, pade_expm, lazy_expm)

Compute bloating factors using forward or backward interpolation.

Input

cs – a continuous system

δ – step size

approx_model – choose the approximation model among "forward" and "backward"

pade_expm – if true, use Pade approximant method to compute the matrix exponential

lazy_expm – if true, compute the matrix exponential in a lazy way suitable for very large systems)

Algorithm

See Frehse et al., CAV'11, SpaceEx: Scalable Verification of Hybrid Systems, Lemma 3.

Note that in the unlikely case that A is invertible, the result can also be obtained directly, as a function of the inverse of A and e^{At} - I.

The matrix P is such that: ϕAabs = P[1:n, 1:n], Phi1Aabsdelta = P[1:n, (n+1):2*n], and Phi2Aabs = P[1:n, (2*n+1):3*n].

source

Discrete-time reachability

Reachability.ReachSets.discr_no_bloat — Function.

discr_no_bloat(cont_sys, δ, pade_expm, lazy_expm)

Discretize a continuous system without bloating of the initial states, suitable for discrete-time reachability.

Input

cont_sys – a continuous system

δ – step size

pade_expm – if true, use Pade approximant method to compute the matrix exponential

lazy_expm – if true, compute the matrix exponential in a lazy way (suitable for very large systems)

Output

A discrete system.

Algorithm

The transformation implemented here is the following:

A -> Phi := exp(A*delta)

U -> V := M*U

X0 -> X0hat := X0

where M corresponds to Phi1(A, delta) in Eq. (8) of SpaceEx: Scalable Verification of Hybrid Systems.

In particular, there is no bloating, i.e. we don't bloat the initial states and dont multiply the input by the step size δ, as required for the dense time case.

source