Systems

Systems

This module provides types to represent systems of affine ODEs in both discrete and continuous time.

Systems

Types of systems

Discrete system

Continuous system

Nondeterministic inputs

Constant nondeterministic inputs

Time-varying nondeterministic inputs

Types of systems

Every system inherits from AbstractSystem.

Reachability.Systems.AbstractSystem — Type.

Abstract type representing a system of affine ODEs.

source

We support the following two concrete types of systems.

Discrete system

A discrete system consists of a matrix representing the system dynamics, a set of initial states, a set of nondeterministic inputs, and a discretization step δ.

Reachability.Systems.DiscreteSystem — Type.

DiscreteSystem <: AbstractSystem

Type that represents a system of discrete-time affine ODEs with nondeterministic inputs,

$x_{k+1} = A x_{k} + u_{k}$

where:

$A$ is a square matrix

$x(0) ∈ \mathcal{X}_0$ and $\mathcal{X}_0$ is a convex set

$u_{k} ∈ \mathcal{U}_{k}$, where $\{\mathcal{U}_{k}\}_k$ is a set-valued sequence defined over $[0, δ], ..., [(N-1)δ, N δ]$ for some $δ>0$

Fields

A – square matrix, possibly of type SparseMatrixExp

X0 – set of initial states

U – nondeterministic inputs

δ – discretization step

Examples

DiscreteSystem(A::Union{AbstractMatrix{Float64}, SparseMatrixExp{Float64}}, X0::LazySet, δ::Float64, U::NonDeterministicInput) – default constructor

DiscreteSystem(A::Union{AbstractMatrix{Float64}, SparseMatrixExp{Float64}}, X0::LazySet, δ::Float64) – constructor with no inputs

DiscreteSystem(A::Union{AbstractMatrix{Float64}, SparseMatrixExp{Float64}}, X0::LazySet, δ::Float64, U::LazySet) – constructor that creates a ConstantNonDeterministicInput

DiscreteSystem(A::Union{AbstractMatrix{Float64}, SparseMatrixExp{Float64}}, X0::LazySet, δ::Float64, U::Vector{<:LazySet}) – constructor that creates a TimeVaryingNonDeterministicInput

source

Reachability.Systems.dim — Method.

dim(S)

Dimension of a discrete system.

Input

S – discrete system

Output

The dimension of the system.

source

Continuous system

A continuous system consists of a matrix representing the system dynamics, a set of initial states, and a set of nondeterministic inputs.

Reachability.Systems.ContinuousSystem — Type.

ContinuousSystem <: AbstractSystem

Type that represents a system of continuous-time affine ODEs with nondeterministic inputs,

$x'(t) = Ax(t) + u(t)$,

where:

$A$ is a square matrix

$x(0) ∈ \mathcal{X}_0$ and $\mathcal{X}_0$ is a convex set

$u(t) ∈ \mathcal{U}(t)$, where $\mathcal{U}(\cdot)$ is a piecewise-constant set-valued function, i.e. we consider that it can be approximated by a possibly time-varying discrete sequence $\{\mathcal{U}_k \}_k$

Fields

A – square matrix

X0 – set of initial states

U – nondeterministic inputs

Examples

ContinuousSystem(A::AbstractMatrix{Float64}, X0::LazySet, U::NonDeterministicInput) – default constructor

ContinuousSystem(A::AbstractMatrix{Float64}, X0::LazySet) – constructor with no inputs

ContinuousSystem(A::AbstractMatrix{Float64}, X0::LazySet, U::LazySet) – constructor that creates a ConstantNonDeterministicInput

ContinuousSystem(A::AbstractMatrix{Float64}, X0::LazySet, U::Vector{<:LazySet}) – constructor that creates a TimeVaryingNonDeterministicInput

source

Reachability.Systems.dim — Method.

dim(S)

Dimension of a continuous system.

Input

S – continuous system

Output

The dimension of the system.

source

Nondeterministic inputs

The above systems may contain nondeterministic inputs, which are wrapped in special types. Every nondeterministic input representation inherits from NonDeterministicInput.

Reachability.Systems.NonDeterministicInput — Type.

Abstract type representing a nondeterministic input. The input can be either constant or time-varying. In both cases it is represented by an iterator.

source

The inputs are closely related to a DiscreteSystem in the sense that for each discrete time step the input set may change. We support iteration through the inputs over time.

Constant nondeterministic inputs

Constant nondeterministic inputs are chosen from a set of values that does not change over time. Note that, while the set is constant, the inputs themselves vary over time.

Reachability.Systems.ConstantNonDeterministicInput — Type.

ConstantNonDeterministicInput <: NonDeterministicInput

Type that represents a constant nondeterministic input.

Fields

U – LazySet

Notes

This type supports iteration with an index number as iterator state. The iteration function next takes and returns a tuple (set, index), where set is the value of the input, represented as a LazySet, and index counts the number of times this iterator was called.

The iterator length is 1, but for convenience next can be called with any index.

Examples

ConstantNonDeterministicInput(U::LazySet) – default constructor

source

Reachability.Systems.next_set — Method.

next_set(inputs, state)

Convenience iteration function that only returns the set.

Input

inputs - nondeterministic inputs wrapper

state - iterator state, i.e., an index

Output

The nondeterministic input set at the given index.

source

Reachability.Systems.next_set — Method.

next_set(inputs)

Convenience iteration function without index that only returns the set.

Input

inputs - constant nondeterministic inputs wrapper

Output

The nondeterministic input set at the given index.

source

Time-varying nondeterministic inputs

Time-varying nondeterministic inputs are chosen from a set of values that changes over time (with each time step).

Reachability.Systems.TimeVaryingNonDeterministicInput — Type.

TimeVaryingNonDeterministicInput <: NonDeterministicInput

Type that represents a time-varying nondeterministic input.

Fields

U – array containing LazySets

Notes

This type supports iteration with an index number as iterator state. The iteration function next takes and returns a tuple (set, index), where set is the value of the input, represented as a LazySet, and index counts the number of times this iterator was called.

The iterator length corresponds to the number of elements in the given array. The index of the input state increases from 1 and corresponds at each time to the array index in the input array.

Examples

TimeVaryingNonDeterministicInput(U::Vector{<:LazySet}) – constructor from a vector of sets

source

Reachability.Systems.next_set — Method.

next_set(inputs, state)

Convenience iteration function that only returns the set.

Input

inputs - nondeterministic inputs wrapper

state - iterator state, i.e., an index

Output

The nondeterministic input set at the given index.

source