Transformations

Transformations

This module provides functions to apply coordinate transformations to Systems using matrix decompositions.

Transformations

Interface

Schur transform

Examples

Interface

This module exports a single function that works as an interface. It dispatches which transformation to apply using a string argument.

Reachability.Transformations.transform — Function.

transform(S; [method])

Interface function that calls the respective transformation function.

Input

S – discrete or continuous system

method – (optional, default: 'schur') transformation method name; valid otions are:

'schur'

Output

A tuple containing:

transformed discrete or continuous system

inverse transformation matrix for reverting the transformation

Notes

The functions that are called in the background should return a the transformed system components A, X0, and U, and also an inverse transformation matrix M.

source

Schur transform

The real Schur decomposition is of the form

\[U^TAU = \begin{pmatrix} T_{11} & T_{12} &\cdots & T_{1b} \\ 0 & T_{22} & \cdots & T_{2b} \\ \vdots & \vdots & \ddots & \vdots \\ 0 & 0 & \cdots & T_{bb} \end{pmatrix}\]

where $T_{ij}$ are 2x2 matrices

Reachability.Transformations.schur_transform — Function.

schur_transform(S)

Applies a Schur transformation to a discrete or continuous system.

Input

S – discrete or continuous system

Output

A tuple containing:

transformed discrete or continuous system

inverse transformation matrix for reverting the transformation

Algorithm

We use Julia's default schurfact function to compute a Schur decomposition of the coefficients matrix $A$.

source

Examples